# CUDA: Quantized matrix matrix multiplication