After https://github.com/ggerganov/llama.cpp/pull/2043 and https://github.com/ggerganov/llama.cpp/pull/2067 I've tried implementing a matrix matrix multiplication kernel using integer intrinsics (currently only q4_0). The results are mixed:

| GPU      | Model    | Test | t/s master | t/s PR | Speedup |
|----------|----------|------|------------|--------|---------|
| RTX 3090 | 7b q4_0  | pp   |       1042 |    924 |    0.89 |
| RTX 3090 | 13b q4_0 | pp   |        677 |    533 |    0.79 |
| RTX 3090 | 33b q4_0 | pp   |        303 |    233 |    0.77 |
| P40      | 7b q4_0  | pp   |        426 |    516 |    1.21 |
| P40      | 13b q4_0 | pp   |        247 |    289 |    1.17 |
| P40      | 33b q4_0 | pp   |        104 |    124 |    1.19 |

On my RTX 3090 the new kernel is slower but on my P40 it's faster. Since matrix matrix multiplications are compute bound I suspect that the reason is that (unlike the P40) the RTX 3090 is capable of executing floating point and integer (i.e. pointer) arithmetic in parallel. So using integer intrinsics instead of floating point operations leaves the floating point hardware underutilized. However, the same GPUs that can execute floating point and integer arithmetic in parallel also have tensor cores which should be much faster anyways so I think that this is not a problem long-term.

Due to the use of shared memory the implementation has gotten rather ugly; using the structs for quantized data in shared memory had terrible performance (most likely due to memory bank conflicts) so the current implementation dissects the data into quants and scales. This also has the unfortunate side effect of tying the allocation of shared memory closely to the quantization type. I very much do not want to implement an entire matrix matrix multiplication kernel for each quantization type but creating a good template could also be tricky; I'll need to think about it some more.