@JohannesGaessler 2023-07-14
After looking into it I think tensor cores should get their own kernel since the use of shared memory will need to be different. Instead I've done some performance optimizations:

| GPU      | Model    | Test | t/s master | t/s PR | Speedup |
|----------|----------|------|------------|--------|---------|
| RTX 3090 | 7b q4_0  | pp   |       1049 |   1050 |    1.00 |
| RTX 3090 | 13b q4_0 | pp   |        676 |    625 |    0.92 |
| RTX 3090 | 33b q4_0 | pp   |        298 |    275 |    0.92 |
| P40      | 7b q4_0  | pp   |        425 |    629 |    1.48 |
| P40      | 13b q4_0 | pp   |        244 |    360 |    1.48 |
| P40      | 33b q4_0 | pp   |        103 |    157 |    1.52 |

I think this performance would be good enough to merge and use as the default; it's possible that the performance for other quantization formats will be worse though.