@JohannesGaessler 2023-07-15
I pushed a template for matrix matrix multiplication. The template accepts three functions:

1. a function that allocates the shared memory to hold the tile for the quantized matrix,
2. a function that loads data from global memory to the aforementioned tile, and
3. a function that does the actual matrix multiplication.