@abc-nix 2023-07-15
While I was testing your new changes (now including your last commit) on RTX 3060 12GB, I found an error while testing tulu-13B-q4_0 (43 layers offloaded) during the PP step that doesn't occur in master.

When I process less than 512 tokens (510t), I get:
`CUDA error 700 at ggml-cuda.cu:3609: an illegal memory access was encountered
`
When I process a bit more (533 tokens), I get this error:
`CUDA error 700 at ggml-cuda.cu:3609: an illegal memory access was encountered`

I tested with pure llama-13b-q4_0 and I don't get this error, so maybe it is just this specific llama finetune. I will download others and test later.

Related to PP speed results, I cannot continue to use tulu, but what I got so far for a prompt 533 tokens long:

|Model|quant|layers|main PP (t/s)|PR PP (t/s)|
|:----|:----|:----|:----|:----|
|7B|q4_K_M|35/35|537.18|487.31|
|7B|q4_0|35/35|505.31|524.49|
|13B|q4_K_M|43/43|289.92|274.90|
|13B|q4_0|43/43|276.84|error|
|30B|q4_K_M|29/63|58.96|57.44|
|30B|q4_0|29/63|57.16|60.16|