@cmp-nct 2023-07-15
This looks so great, I can't wait to get rid of cuBLAS entirely. 
Maybe once a 8_0 variant is available it could already be used as a full cuBLAS replacement without sacrificing quality in as good as all scenarios? 
After all, for cublas we currently live-convert/dequantize the inputs anyway to 32/16. Doing that down to just 8 would be quite an improvement already.