@JohannesGaessler 2023-07-15
The out-of-bounds memory accesses should be fixed now.