@JohannesGaessler 2023-07-16
I implemented all of the older quantization formats and rebased onto master. These are the results:

| GPU      | Model   | Test | t/s master | t/s PR | Speedup |
|----------|---------|------|------------|--------|---------|
| RTX 3090 | 7b q4_0 | pp   |       1292 |   1266 |    0.98 |
| RTX 3090 | 7b q4_1 | pp   |       1283 |   1263 |    0.98 |
| RTX 3090 | 7b q5_0 | pp   |       1277 |    647 |    0.51 |
| RTX 3090 | 7b q5_1 | pp   |       1269 |    634 |    0.50 |
| RTX 3090 | 7b q8_0 | pp   |       1273 |   1039 |    0.82 |
| P40      | 7b q4_0 | pp   |        461 |    627 |    1.36 |
| P40      | 7b q4_1 | pp   |        459 |    368 |    0.80 |
| P40      | 7b q5_0 | pp   |        460 |    245 |    0.53 |
| P40      | 7b q5_1 | pp   |        451 |    254 |    0.56 |
| P40      | 7b q8_0 | pp   |        453 |    532 |    1.17 |

Using cuBLAS the performance is very consistent. The performance of the new kernels varies a lot depending on quantization format. I think the fundamental reason is that for matrix matrix multiplication you are compute bound and I/O is much less important. So it can be faster to dequantize the entire weight matrix once and then work with f16/f32 values which you can multiply directly (and which is comparatively fast on GPUs) than it is to do multiple integer/logical operations on the quantized data to get the result. 

Notably, the performance for q5_0 and q5_1 is bad, presumably because the 5th bits are ordered in an inconvenient way that requires 4 bit shifts, bit-wise ANDs, and bit-wise ORs. This could be reduced to a single bit shift, bit-wise AND, and bit-wise OR (could be done when the weights are loaded into VRAM if the different bit order is bad for CPU performance). The performance of q4_1 on the P40 is also relatively bad, the problem being that for good performance some f16 calculations are necessary which have bad performance on the P40 (a f32 workaround is used instead).

Overall I expect the performance for q2_k, q3_k, and q5_k to not be good due to the large number of operations per data value that will be necessary.

Caveats: the performance can probably still be optimized a lot. In particular I think there is still potential to optimize memory bank conflicts and tile sizes. For Ampere or newer it's also possible to utilize asynchronous data loading.

@ggerganov @slaren What is your judgement regarding prompt processing speed vs. VRAM usage? I personally would prefer lower VRAM usage as the default because I find prompt processing to be fast enough either way. Also my findings may apply to a lesser extent to CPU matrix matrix multiplication as well. CPUs generally have comparatively fast integer arithmetic though and I think the memory bandwidth is a much more severe bottleneck for CPUs than for GPUs.