@JohannesGaessler 2023-07-28
I pushed implementations for k-quants and did some tests:

| GPU      | Model     | Test | t/s master | t/s PR | Speedup | VRAM master [MiB] | VRAM PR [MiB] | VRAM saved [MiB] |
|----------|-----------|------|------------|--------|---------|-------------------|---------------|------------------|
| RTX 3090 | 7b q4_0   | pp   |       1302 |   1265 |    0.97 |              5454 |          4754 |              700 |
| RTX 3090 | 13b q4_0  | pp   |        775 |    711 |    0.92 |              9306 |          8336 |              970 |
| RTX 3090 | 33b q4_0  | pp   |        324 |    302 |    0.93 |             20418 |         18988 |             1430 |
| RTX 3090 | 7b q4_1   | pp   |       1293 |   1209 |    0.94 |              5678 |          4978 |              700 |
| RTX 3090 | 7b q5_0   | pp   |       1274 |    629 |    0.49 |              6126 |          5426 |              700 |
| RTX 3090 | 7b q5_1   | pp   |       1266 |    621 |    0.49 |              6510 |          5810 |              700 |
| RTX 3090 | 7b q8_0   | pp   |       1267 |   1003 |    0.79 |              8430 |          7730 |              700 |
| RTX 3090 | 7b q2_k   | pp   |       1276 |    502 |    0.39 |              4528 |          3828 |              700 |
| RTX 3090 | 7b q3_k_s | pp   |       1286 |    342 |    0.27 |              4654 |          3954 |              700 |
| RTX 3090 | 7b q4_k_s | pp   |       1271 |    485 |    0.38 |              5486 |          4786 |              700 |
| RTX 3090 | 7b q5_k_s | pp   |       1283 |    406 |    0.32 |              6126 |          5426 |              700 |
| RTX 3090 | 7b q6_k   | pp   |       1275 |    441 |    0.35 |              6958 |          6258 |              700 |
| P40      | 7b q4_0   | pp   |        467 |    616 |    1.32 |              5420 |          4720 |              700 |
| P40      | 13b q4_0  | pp   |        259 |    344 |    1.33 |              9272 |          8302 |              970 |
| P40      | 33b q4_0  | pp   |        107 |    146 |    1.36 |             20384 |         18954 |             1430 |
| P40      | 7b q4_1   | pp   |        468 |    354 |    0.76 |              5644 |          4944 |              700 |
| P40      | 7b q5_0   | pp   |        459 |    247 |    0.54 |              6074 |          5374 |              700 |
| P40      | 7b q5_1   | pp   |        446 |    252 |    0.57 |              6466 |          5766 |              700 |
| P40      | 7b q8_0   | pp   |        452 |    526 |    1.16 |              8426 |          7726 |              700 |
| P40      | 7b q2_k   | pp   |        455 |    177 |    0.39 |              4494 |          3794 |              700 |
| P40      | 7b q3_k_s | pp   |        438 |    113 |    0.26 |              4620 |          3920 |              700 |
| P40      | 7b q4_k_s | pp   |        446 |    208 |    0.47 |              5452 |          4752 |              700 |
| P40      | 7b q5_k_s | pp   |        461 |    153 |    0.33 |              6092 |          5392 |              700 |
| P40      | 7b q6_k   | pp   |        462 |    161 |    0.35 |              6924 |          6224 |              700 |

VRAM usage reduction does not depend on quantization, only on model size (and which layers are being offloaded). The k-quant performance is very bad, I'll try to at least get the low-hanging fruits this weekend.