@JohannesGaessler 2023-07-29
I did some performance optimizations:

| GPU      | Model     | Test | t/s master | t/s PR | Speedup |
|----------|-----------|------|------------|--------|---------|
| RTX 3090 | 7b q4_0   | pp   |       1302 |   1354 |    1.04 |
| RTX 3090 | 13b q4_0  | pp   |        775 |    766 |    0.99 |
| RTX 3090 | 33b q4_0  | pp   |        324 |    323 |       1 |
| RTX 3090 | 7b q4_1   | pp   |       1293 |   1293 |       1 |
| RTX 3090 | 7b q5_0   | pp   |       1274 |    972 |    0.76 |
| RTX 3090 | 7b q5_1   | pp   |       1266 |    895 |    0.71 |
| RTX 3090 | 7b q8_0   | pp   |       1267 |   1066 |    0.84 |
| RTX 3090 | 7b q2_k   | pp   |       1276 |    642 |     0.5 |
| RTX 3090 | 7b q3_k_s | pp   |       1286 |    473 |    0.37 |
| RTX 3090 | 7b q4_k_s | pp   |       1271 |    833 |    0.66 |
| RTX 3090 | 7b q5_k_s | pp   |       1283 |    518 |     0.4 |
| RTX 3090 | 7b q6_k   | pp   |       1275 |    587 |    0.46 |
| P40      | 7b q4_0   | pp   |        467 |    623 |    1.33 |
| P40      | 13b q4_0  | pp   |        259 |    347 |    1.34 |
| P40      | 33b q4_0  | pp   |        107 |    148 |    1.38 |
| P40      | 7b q4_1   | pp   |        468 |    359 |    0.77 |
| P40      | 7b q5_0   | pp   |        459 |    247 |    0.54 |
| P40      | 7b q5_1   | pp   |        446 |    254 |    0.57 |
| P40      | 7b q8_0   | pp   |        452 |    527 |    1.17 |
| P40      | 7b q2_k   | pp   |        455 |    209 |    0.46 |
| P40      | 7b q3_k_s | pp   |        438 |    120 |    0.27 |
| P40      | 7b q4_k_s | pp   |        446 |    228 |    0.51 |
| P40      | 7b q5_k_s | pp   |        461 |    201 |    0.44 |
| P40      | 7b q6_k   | pp   |        462 |    161 |    0.35 |

Part of this is that I added a compile option `LLAMA_CUDA_MMQ_Y` which can be used to control tile size; on my RTX 3090 a tile size of 128 is faster than a tile size of 64. However, 128 is much slower on my P40 so I'm keeping the default of 64.